Refactor and javadoc.
