+++ src/main/java/es/esky/role/users/domain/validator/Uppercase.java
@@ -34,10 +34,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Retention(RUNTIME)
 @Constraint(validatedBy = UppercaseValidator.class)
 public @interface Uppercase {
-
-    String message() default "{es.esky.role.user.validation.constraints.upper_case}";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
+	String message() default "{es.esky.role.user.validation.constraints.upper_case}";
+	Class<?>[] groups() default {};
+	Class<? extends Payload>[] payload() default {};
 }
